backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pymongo>=4.6.0
shodan>=1.31.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
